moves fetchlocaldata from hook to helper function
