+++ play/src/lib/HelperFunctions.tsx
@@ -75,3 +75,20 @@ export const checkForSubmittedAnswerOnRejoin = (
   }
   return { selectedAnswerIndex, isSubmitted };
 }
+
+export const fetchLocalData = () => {
+  const pregameModel = window.localStorage.getItem('rightOn');
+  if (isNullOrUndefined(pregameModel)) return null;
+
+  const parsedPregameModel = JSON.parse(pregameModel);
+  // checks for invalid data in pregameModel, returns null if found
+  if (
+    isNullOrUndefined(parsedPregameModel.gameSessionId) ||
+    isNullOrUndefined(parsedPregameModel.teamId) ||
+    isNullOrUndefined(parsedPregameModel.teamMemberId) ||
+    isNullOrUndefined(parsedPregameModel.selectedAvatar)
+  )
+    return null;
+  // passes validated pregameModel to GameInProgressContainer
+  return parsedPregameModel;
+}
